End Scene
Created an end condition and an end scene. The end scene displays the player ID and the score. Score and id are being saved in the game. End condition is if proficiency reaches 40, meaning the player is proficient in both levels. Left to decide -- end condition if never reaches proficiency